+++ sdk/README.md
@@ -43,10 +43,10 @@ const tools = [createRecordTool, updateRecordTool];
 ### Using with Mastra
 
 ```typescript
-import { createActionTool, updateActionTool } from "@amp-labs/ai/mastra";
+import { createRecordTool, updateRecordTool } from "@amp-labs/ai/mastra";
 
 // Use in your Mastra workflow
-const tools = [createActionTool, updateActionTool];
+const tools = [createRecordTool, updateRecordTool];
 ```
 `
 
@@ -55,7 +55,7 @@ const tools = [createActionTool, updateActionTool];
 If you encounter issues, ensure that:
 
 1. All required environment variables are set correctly
-2. You have an active Ampersand account with the necessary permissions
-3. The SaaS tool connection has been properly established
+2. You have an active Ampersand account with the necessary permissions & scopes
+3. The SaaS tool (say HubSpot) connection (with OAuth or apiKey) with your Agent is properly established & is still active. 
 
 For further assistance, join our [Discord community](https://discord.gg/BWP4BpKHvf).
