+++ sdk/lib/adapters/ampersand/types/index.ts
@@ -1,0 +1,37 @@
+// Type definitions for the parameters
+export type CreateParams = {
+  provider: string;
+  objectName: string;
+  type: "create";
+  record: Record<string, any>;
+  groupRef: string;
+  associations?: Array<{
+    to: { id: string };
+    types: Array<{
+      associationCategory: string;
+      associationTypeId: number;
+    }>;
+  }>;
+};
+
+export type UpdateParams = {
+  provider: string;
+  objectName: string;
+  type: "update";
+  record: Record<string, any>;
+  groupRef: string;
+  associations?: Array<{
+    to: { id: string };
+    types: Array<{
+      associationCategory: string;
+      associationTypeId: number;
+    }>;
+  }>;
+};
+
+export type WriteResponse = {
+  success: boolean;
+  status: string;
+  recordId: string;
+  response: any;
+};
